move and rename the adapters
