+++ app/src/main/java/com/example/ohthmhyh/adapters/FriendRequestListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.ohthmhyh;
+package com.example.ohthmhyh.adapters;
 
 import android.content.Context;
 import android.util.Log;
@@ -9,6 +9,7 @@ import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.TextView;
 
+import com.example.ohthmhyh.R;
 import com.example.ohthmhyh.database.DatabaseAdapter;
 
 import java.util.ArrayList;
