+++ app/src/main/java/com/example/ohthmhyh/adapters/FriendsListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.ohthmhyh;
+package com.example.ohthmhyh.adapters;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -7,6 +7,7 @@ import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.TextView;
 
+import com.example.ohthmhyh.R;
 import com.example.ohthmhyh.database.DatabaseAdapter;
 
 import java.util.ArrayList;
