+++ app/src/main/java/com/example/ohthmhyh/fragments/HabitsFragment.java
@@ -18,7 +18,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 
-import com.example.ohthmhyh.CustomAdapterHF;
+import com.example.ohthmhyh.adapters.HabitRecyclerViewAdapter;
 import com.example.ohthmhyh.activities.UpdateHabitActivity;
 import com.example.ohthmhyh.database.DatabaseAdapter;
 import com.example.ohthmhyh.entities.Habit;
@@ -29,14 +29,14 @@ import com.example.ohthmhyh.helpers.TransportableTouchHelper;
 /**
  * A simple {@link Fragment} subclass.
  */
-public class HabitsFragment extends Fragment implements CustomAdapterHF.OntouchListener {
+public class HabitsFragment extends Fragment implements HabitRecyclerViewAdapter.OntouchListener {
 
     public static final String ARG_RETURNED_HABIT = "returned_habit_arg";
 
     private int chosenHabitIndex = -1;
     private ActivityResultLauncher<Intent> resultLauncher;
     private HabitList habitList;
-    private CustomAdapterHF adapter;
+    private HabitRecyclerViewAdapter adapter;
     private DatabaseAdapter databaseAdapter;
 
     public HabitsFragment(){
@@ -71,7 +71,7 @@ public class HabitsFragment extends Fragment implements CustomAdapterHF.OntouchL
 
                 recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
                 recyclerView.setHasFixedSize(true);
-                adapter = new CustomAdapterHF(getContext(), HabitsFragment.this, habitList);
+                adapter = new HabitRecyclerViewAdapter(getContext(), HabitsFragment.this, habitList);
                 ItemTouchHelper.Callback callback = new TransportableTouchHelper(adapter);
                 ItemTouchHelper itemTouchHelper = new ItemTouchHelper(callback);
                 adapter.setTouchhelper(itemTouchHelper);
